sound: msm: qdsp6v2: revert most joan-related code to release version
The G7 One used code from the LG G7 ThinQ kernel. This wasn't a problem
for phoenix, which didn't have specific hardware that the V-series
devices do. But when compiling for joan, this became an issue.

Seeing as the code in question is unused otherwise, we can blanket copy
back joan code into relevant sections.

Change-Id: Iad61e4d596e386977461776441ed671662a21933
Signed-off-by: Han Sol Jin <hansol@hansol.ca>